+++ frontend/src/components/index.js
@@ -8,3 +8,4 @@ export { Sidebar } from "./Sidebar";
 export { DetailsList } from "./DetailsList";
 export { Table } from "./Table";
 export { GDXAccordion } from "./GDXAccordion";
+export { ReportSelect } from "./ReportSelect";