+++ frontend/src/pages/Reports/Report/index.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+/**
+ * This reusable component renders the projects component
+ *
+ * @param   {object}             query React query that contains all projects
+ * @returns {React.ReactElement}       The project component
+ */
+
+export const Report = () => {
+  return (
+    <>
+      <p>Hello World</p>
+    </>
+  );
+};