+++ frontend/src/routes/subRoutes/reportRoutes.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Route } from "react-router-dom";
+import { Reports } from "../../pages";
+const reportRoutes = [<Route key="report" path="reports" element={<Reports />} />];
+
+export default reportRoutes;