+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## Developer Changelog
 
+### August 17, 2022 (DESCW-457)
+- Frontend
+  - Added Reports navigation and report selector
+
 ### August 17, 2020 (DESCW-538)
 
 - Frontend
   - Resources view modal
 - Backend
-
   - Added controller, model, route, validator for project status summary
 
   - Requires
@@ -15,7 +18,6 @@
   - Added controller, model, route, validator for project status summary
 
 ### August 16, 2022 (DESCW-505)
-
 - Frontend
 - Added test for GDX Accordion